fix assert reference
