--- packages/project/src/assert/assertion-error.ts
@@ -1,7 +1,0 @@
-export class AssertionError extends Error {
-  constructor(message: string) {
-    super(message);
-
-    Object.setPrototypeOf(this, new.target.prototype);
-  }
-}
