+++ packages/project/test/assertion-error.test.ts
@@ -1,4 +1,4 @@
-import { AssertionError } from '../src/assert/assertion-error';
+import { AssertionError } from '../src';
 import * as expect from 'expect';
 
 describe('AssertionError', () => {
